main: rename output*.[ch] to writer*.[ch]
