+++ source.mak
@@ -30,7 +30,6 @@ MAIN_HEADS =			\
 	main/nestlevel.h	\
 	main/objpool.h		\
 	main/options.h		\
-	main/output.h		\
 	main/param.h		\
 	main/parse.h		\
 	main/parsers.h		\
@@ -45,6 +44,7 @@ MAIN_HEADS =			\
 	main/strlist.h		\
 	main/types.h		\
 	main/vstring.h		\
+	main/writer.h		\
 	main/xtag.h		\
 	\
 	$(MIO_HEADS)
@@ -68,11 +68,6 @@ MAIN_SRCS =				\
 	main/nestlevel.c		\
 	main/objpool.c			\
 	main/options.c			\
-	main/output.c			\
-	main/output-etags.c		\
-	main/output-ctags.c		\
-	main/output-json.c		\
-	main/output-xref.c		\
 	main/param.c			\
 	main/parse.c			\
 	main/pcoproc.c			\
@@ -85,6 +80,11 @@ MAIN_SRCS =				\
 	main/sort.c			\
 	main/strlist.c			\
 	main/vstring.c			\
+	main/writer.c			\
+	main/writer-etags.c		\
+	main/writer-ctags.c		\
+	main/writer-json.c		\
+	main/writer-xref.c		\
 	main/xtag.c			\
 	\
 	$(REPOINFO_SRCS) \
